simple socket connection

+++ SprintBootAppWithSQL/src/main/java/com/example/SprintBootAppWithSQL/props/jwtProperties.java
@@ -25,7 +25,7 @@ public class jwtProperties {
     @Getter
     @Setter
     @Value("${jwt.expiry}")
-    private Integer expiry;
+    private Long expiry;
     @Getter
     @Setter
     @Value("${jwt.refresh.token.expiry}")
